Add an alternative implementation rasterises per-edge
This is based on the libart/pathfinder/font-rs approach and
differs from the original implementation which rasterises
per-triangle, to give a ~2x improvement on highly connected meshes.
It involves an initial tessellation step to produce a list of triangles
and then edges from a rectangular grid.

Cannot currently clip to a smaller viewport than the entire extent of
the mesh to be rasterized.